Fix: Use nextjs file naming convention to rename files and folders
